apps/rating-engine/pyproject.toml
[project]
name = "rating-engine"
version = "0.1.0"
description = "OpenSkill rating calculation engine for Riichi Mahjong League"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "openskill>=5.0.0",
    "pydantic>=2.5.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.16.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "postgrest>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rating_engine"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
# Conservative enhancement - add quality rules gradually
select = [
    "E",    # pycodestyle errors
    "F",    # Pyflakes  
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear (common bugs)
    "C4",   # flake8-comprehensions (better comprehensions)
    "PERF", # Performance optimizations
]

ignore = [
    # Allow some flexibility for working code
    "B904",   # raise-without-from (can be addressed later)
]

# Conservative fixable settings
fixable = ["I", "UP", "C4", "PERF"]  # Safe auto-fixes only
unfixable = ["F", "E", "B"]         # Manual review needed

[tool.ruff.lint.per-file-ignores]
# Scripts can be more permissive
"scripts/*" = ["B", "PERF"]
# Tests have different patterns  
"tests/*" = ["B"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--verbose", 
    "--tb=short"
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.pytest.markers]
slow = "marks tests as slow (deselect with '-m \"not slow\"')"
integration = "marks tests as integration tests"
unit = "marks tests as unit tests"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.12.3",
]
